- created structures listed on page load - new structures appear on the list after creation - wallet (still 1) to assign to the group selected from the observed wallets - moved useObservedWallets to shared loaders file
