+++ src/components/structures/details/index.tsx
@@ -1,28 +1,31 @@
 import { type Signal, component$ } from "@builder.io/qwik";
-import {type Structure, StructureBalance} from "~/interface/structure/Structure";
+import {
+  type Structure,
+  StructureBalance,
+} from "~/interface/structure/Structure";
 
 interface SelectedStructureProps {
-    selectedStructure: Signal<StructureBalance| null>;
-    isDeleteModalopen: Signal<boolean>;
+  selectedStructure: Signal<StructureBalance | null>;
+  isDeleteModalopen: Signal<boolean>;
 }
 
 export const SelectedStructureDetails = component$<SelectedStructureProps>(
-    ({ selectedStructure, isDeleteModalopen }) => {
-        if (!selectedStructure.value) return <></>;
-        return (
-            <div class="text-white">
-                <div class="mb-2 flex w-full items-center justify-between rounded p-2">
-                    <span>{selectedStructure.value.structure.name}</span>
-                    <button
-                        class="cursor-pointer rounded bg-red-500 p-2 text-white"
-                        onClick$={() => {
-                            isDeleteModalopen.value = !isDeleteModalopen.value;
-                        }}
-                    >
-                        Delete
-                    </button>
-                </div>
-            </div>
-        );
-    },
+  ({ selectedStructure, isDeleteModalopen }) => {
+    if (!selectedStructure.value) return <></>;
+    return (
+      <div class="text-white">
+        <div class="mb-2 flex w-full items-center justify-between rounded p-2">
+          <span>{selectedStructure.value.structure.name}</span>
+          <button
+            class="cursor-pointer rounded bg-red-500 p-2 text-white"
+            onClick$={() => {
+              isDeleteModalopen.value = !isDeleteModalopen.value;
+            }}
+          >
+            Delete
+          </button>
+        </div>
+      </div>
+    );
+  },
 );
