+++ src/routes/app/wallets/index.tsx
@@ -24,6 +24,8 @@ import { isAddress } from "viem";
 // import ImgSearch from "../../../../../public/images/svg/search.svg?jsx";
 import ImgArrowDown from "/public/images/arrowDown.svg?jsx";
 import ImgI from "/public/images/svg/i.svg?jsx";
+import { useObservedWallets } from "~/routes/shared/loaders";
+export { useObservedWallets } from "~/routes/shared/loaders";
 
 export const useAddWallet = routeAction$(
   async (data, requestEvent) => {
@@ -142,82 +144,83 @@ export const useRemoveWallet = routeAction$(
   }),
 );
 
-export const useObservedWallets = routeLoader$(async (requestEvent) => {
-  const db = await connectToDB(requestEvent.env);
-
-  const cookie = requestEvent.cookie.get("accessToken");
-  if (!cookie) {
-    throw new Error("No cookie found");
-  }
-  const { userId } = jwt.decode(cookie.value) as JwtPayload;
-
-  const [result]: any = await db.query(
-    `SELECT ->observes_wallet.out FROM ${userId};`,
-  );
-  if (!result) throw new Error("No observed wallets");
-  const observedWalletsQueryResult = result[0]["->observes_wallet"].out;
-
-  const observedWallets: any[] = [];
-  for (const observedWallet of observedWalletsQueryResult) {
-    const [wallet] = await db.select<Wallet>(`${observedWallet}`);
-    const nativeBalance = await publicClient.getBalance({
-      address: wallet.address as `0x${string}`,
-      blockTag: "safe",
-    });
-    await db.query(
-      `UPDATE ${observedWallet} SET nativeBalance = '${nativeBalance}';`,
-    );
-
-    const walletTokensBalances: WalletTokensBalances = {
-      wallet: {
-        id: wallet.id,
-        name: wallet.name,
-        chainId: wallet.chainId,
-        address: wallet.address,
-        nativeBalance: nativeBalance,
-      },
-      tokens: [],
-    };
-
-    // For each token update balance
-    const tokens = await db.select<Token>("token");
-    for (const token of tokens) {
-      const readBalance = await publicClient.readContract({
-        address: token.address as `0x${string}`,
-        abi: contractABI,
-        functionName: "balanceOf",
-        args: [wallet.address as `0x${string}`],
-      });
-
-      // Certain balance which shall be updated
-      const [[balanceToUpdate]]: any = await db.query(
-        `SELECT * FROM balance WHERE ->(for_wallet WHERE out = '${wallet.id}') AND ->(for_token WHERE out = '${token.id}');`,
-      );
-
-      await db.update<Balance>(`${balanceToUpdate.id}`, {
-        value: readBalance.toString(),
-      });
-
-      const formattedBalance = formatTokenBalance(
-        readBalance.toString(),
-        token.decimals,
-      );
-
-      if (readBalance !== BigInt(0) && formattedBalance !== "0.000") {
-        // Add the token to the wallet object
-        walletTokensBalances.tokens.push({
-          id: token.id,
-          name: token.name,
-          symbol: token.symbol,
-          decimals: token.decimals,
-          balance: formattedBalance,
-        });
-      }
-    }
-    observedWallets.push(walletTokensBalances);
-  }
-  return observedWallets;
-});
+// export const useObservedWallets = routeLoader$(async (requestEvent) => {
+//   const db = await connectToDB(requestEvent.env);
+//
+//   const cookie = requestEvent.cookie.get("accessToken");
+//   if (!cookie) {
+//     throw new Error("No cookie found");
+//   }
+//   const { userId } = jwt.decode(cookie.value) as JwtPayload;
+//
+//   const [result]: any = await db.query(
+//     `SELECT ->observes_wallet.out FROM ${userId};`,
+//   );
+//   if (!result) throw new Error("No observed wallets");
+//   const observedWalletsQueryResult = result[0]["->observes_wallet"].out;
+//
+//   const observedWallets: any[] = [];
+//   for (const observedWallet of observedWalletsQueryResult) {
+//     const [wallet] = await db.select<Wallet>(`${observedWallet}`);
+//     const nativeBalance = await publicClient.getBalance({
+//       address: wallet.address as `0x${string}`,
+//       blockTag: "safe",
+//     });
+//     await db.query(
+//       `UPDATE ${observedWallet} SET nativeBalance = '${nativeBalance}';`,
+//     );
+//
+//     const walletTokensBalances: WalletTokensBalances = {
+//       wallet: {
+//         id: wallet.id,
+//         name: wallet.name,
+//         chainId: wallet.chainId,
+//         address: wallet.address,
+//         nativeBalance: nativeBalance,
+//       },
+//       tokens: [],
+//     };
+//
+//     // For each token update balance
+//     const tokens = await db.select<Token>("token");
+//     for (const token of tokens) {
+//       const readBalance = await publicClient.readContract({
+//         address: token.address as `0x${string}`,
+//         abi: contractABI,
+//         functionName: "balanceOf",
+//         args: [wallet.address as `0x${string}`],
+//       });
+//
+//       // Certain balance which shall be updated
+//       const [[balanceToUpdate]]: any = await db.query(
+//         `SELECT * FROM balance WHERE ->(for_wallet WHERE out = '${wallet.id}') AND ->(for_token WHERE out = '${token.id}');`,
+//       );
+//
+//       await db.update<Balance>(`${balanceToUpdate.id}`, {
+//         value: readBalance.toString(),
+//       });
+//
+//       const formattedBalance = formatTokenBalance(
+//         readBalance.toString(),
+//         token.decimals,
+//       );
+//
+//       if (readBalance !== BigInt(0) && formattedBalance !== "0.000") {
+//         // Add the token to the wallet object
+//         walletTokensBalances.tokens.push({
+//           id: token.id,
+//           name: token.name,
+//           symbol: token.symbol,
+//           decimals: token.decimals,
+//           balance: formattedBalance,
+//         });
+//       }
+//     }
+//     observedWallets.push(walletTokensBalances);
+//   }
+//   console.log(observedWallets)
+//   return observedWallets;
+// });
 
 export default component$(() => {
   const addWalletAction = useAddWallet();
